web_app/requirements.txt
Flask
google-cloud-datastore
google-auth
requests

google-api-python-client
google-auth-oauthlib
google-auth-httplib2

fitbit